url config done

+++ courses/views/course_crud.py
@@ -2,22 +2,30 @@ from rest_framework.response import Response
 from rest_framework.decorators import api_view
 from rest_framework import status
 from courses.models import Course
-from courses.serializers import CourseSerializer
+from courses.serializers import CourseSerializers
 
-@api_view(['GET', 'PUT', 'DELETE'])
+@api_view(['GET', 'PUT', 'PATCH', 'DELETE'])
 def Course_crud(request, pk):
     course = Course.objects.get(pk=pk)
     if request.method == 'GET':
-        serializer = CourseSerializer(course)
-        return Response(serializer.data, status=status.HTTP_200_CREATED)
+        serializer = CourseSerializers(course)
+        return Response(serializer.data, status=status.HTTP_200_OK)
     elif request.method == 'PUT':
-        serializer = CourseSerializer(course, data=request.data)
+        serializer = CourseSerializers(request.data, instance = course)
         if serializer.is_valid():
             serializer.save()
-            return Response(serializer.data,status=status.HTTP_200_CREATED)
-        return Response(serializer.error, status=status.HTTP_200_CREATED)
+            return Response(serializer.data, status=status.HTTP_200_OK)
+        return Response(status=status.HTTP_400_BAD_REQUEST)
+    elif request.method == 'PATCH':
+        serializer = CourseSerializers(request.data, instance = course, partial = True)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_200_OK)
+        return Response(status=status.HTTP_400_BAD_REQUEST)
     elif request.method == 'DELETE':
-        course.delete()
-        return Response({'message':'Course Deleted'})
+        if course is not None:
+            course.delete()
+            return Response({'message':'Course Deleted'})
+        return Response({'message':'Course Not Found'})
     else:
-        return Response({'message':'Not Found'})
+        return Response(status=status.HTTP_400_BAD_REQUEST)
